+++ extension-androidauto/src/main/java/com/mapbox/maps/extension/androidauto/MapboxCarMap.kt
@@ -3,10 +3,13 @@ package com.mapbox.maps.extension.androidauto
 import android.graphics.Rect
 import androidx.car.app.AppManager
 import androidx.car.app.CarContext
+import androidx.car.app.Session
 import androidx.lifecycle.Lifecycle
 import com.mapbox.maps.EdgeInsets
 import com.mapbox.maps.MapInitOptions
 import com.mapbox.maps.MapboxExperimental
+import com.mapbox.maps.extension.androidauto.internal.defaultMapInitOptions
+import com.mapbox.maps.extension.androidauto.internal.verifyCarContext
 
 /**
  * This is the main entry point for controlling the Mapbox car map surface.
@@ -27,13 +30,16 @@ import com.mapbox.maps.MapboxExperimental
  * The internals of this class use [AppManager.setSurfaceCallback], which assumes there is a single
  * surface callback. Do not use setSurfaceCallback, and do not create multiple instances of
  * [MapboxCarMap].
- *
- * @param mapInitOptions to initialize the [MapboxCarMapSurface]
  */
 @MapboxExperimental
-class MapboxCarMap(
-  private val mapInitOptions: MapInitOptions
-) {
+class MapboxCarMap {
+
+  /**
+   * The initial set of options used to [setup] the map.
+   */
+  lateinit var mapInitOptions: MapInitOptions
+    private set
+
   /**
    * Accessor for the carContext provided to the MapInitOptions. This makes it easier to create
    * screens with the MapboxCarMap in the constructor.
@@ -44,18 +50,22 @@ class MapboxCarMap(
    * The carContext can also be found in the [MapboxCarMapObserver] callbacks. Make sure to
    * call [MapboxCarMap.clearObservers] when your car session is destroyed.
    */
-  val carContext: CarContext = mapInitOptions.context as? CarContext
-    ?: throw IllegalArgumentException("You must construct a MapboxCarMap with the CarContext")
+  lateinit var carContext: CarContext
+    private set
 
   private val carMapSurfaceOwner = CarMapSurfaceOwner()
-  private val carMapSurfaceCallback = CarMapSurfaceCallback(
-    carContext,
-    carMapSurfaceOwner,
-    mapInitOptions
-  )
 
-  init {
-    carMapSurfaceCallback.onBind()
+  /**
+   * @param session Android Auto session that will show the Mapbox Map
+   * @param mapInitOptions optional parameter to initialize the [MapboxCarMapSurface]
+   */
+  fun setup(
+    session: Session,
+    mapInitOptions: MapInitOptions = session.carContext.defaultMapInitOptions()
+  ) = apply {
+    mapInitOptions.verifyCarContext()
+    this.mapInitOptions = mapInitOptions
+    carMapSurfaceOwner.setup(carContext, mapInitOptions)
   }
 
   /**